Extract completion service helper.
