+++ src/OmniSharp.Roslyn.CSharp/Services/Completion/CompletionListBuilder.cs
@@ -1,0 +1,89 @@
+﻿#nullable enable
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Completion;
+using Microsoft.CodeAnalysis.Tags;
+using Microsoft.CodeAnalysis.Text;
+using OmniSharp.Models;
+using OmniSharp.Models.v1.Completion;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using CompletionItem = OmniSharp.Models.v1.Completion.CompletionItem;
+using CSharpCompletionList = Microsoft.CodeAnalysis.Completion.CompletionList;
+using CSharpCompletionService = Microsoft.CodeAnalysis.Completion.CompletionService;
+
+namespace OmniSharp.Roslyn.CSharp.Services.Completion
+{
+    internal static partial class CompletionListBuilder
+    {
+        private static readonly Dictionary<string, CompletionItemKind> s_roslynTagToCompletionItemKind = new()
+        {
+            { WellKnownTags.Public, CompletionItemKind.Keyword },
+            { WellKnownTags.Protected, CompletionItemKind.Keyword },
+            { WellKnownTags.Private, CompletionItemKind.Keyword },
+            { WellKnownTags.Internal, CompletionItemKind.Keyword },
+            { WellKnownTags.File, CompletionItemKind.File },
+            { WellKnownTags.Project, CompletionItemKind.File },
+            { WellKnownTags.Folder, CompletionItemKind.Folder },
+            { WellKnownTags.Assembly, CompletionItemKind.File },
+            { WellKnownTags.Class, CompletionItemKind.Class },
+            { WellKnownTags.Constant, CompletionItemKind.Constant },
+            { WellKnownTags.Delegate, CompletionItemKind.Function },
+            { WellKnownTags.Enum, CompletionItemKind.Enum },
+            { WellKnownTags.EnumMember, CompletionItemKind.EnumMember },
+            { WellKnownTags.Event, CompletionItemKind.Event },
+            { WellKnownTags.ExtensionMethod, CompletionItemKind.Method },
+            { WellKnownTags.Field, CompletionItemKind.Field },
+            { WellKnownTags.Interface, CompletionItemKind.Interface },
+            { WellKnownTags.Intrinsic, CompletionItemKind.Text },
+            { WellKnownTags.Keyword, CompletionItemKind.Keyword },
+            { WellKnownTags.Label, CompletionItemKind.Text },
+            { WellKnownTags.Local, CompletionItemKind.Variable },
+            { WellKnownTags.Namespace, CompletionItemKind.Module },
+            { WellKnownTags.Method, CompletionItemKind.Method },
+            { WellKnownTags.Module, CompletionItemKind.Module },
+            { WellKnownTags.Operator, CompletionItemKind.Operator },
+            { WellKnownTags.Parameter, CompletionItemKind.Variable },
+            { WellKnownTags.Property, CompletionItemKind.Property },
+            { WellKnownTags.RangeVariable, CompletionItemKind.Variable },
+            { WellKnownTags.Reference, CompletionItemKind.Reference },
+            { WellKnownTags.Structure, CompletionItemKind.Struct },
+            { WellKnownTags.TypeParameter, CompletionItemKind.TypeParameter },
+            { WellKnownTags.Snippet, CompletionItemKind.Snippet },
+            { WellKnownTags.Error, CompletionItemKind.Text },
+            { WellKnownTags.Warning, CompletionItemKind.Text },
+        };
+
+        // VS has a more complex concept of a commit mode vs suggestion mode for intellisense.
+        // LSP doesn't have this, so mock it as best we can by removing space ` ` from the list
+        // of commit characters if we're in suggestion mode.
+        private static readonly IReadOnlyList<char> DefaultRulesWithoutSpace = CompletionRules.Default.DefaultCommitCharacters.Where(c => c != ' ').ToList();
+
+        internal static async Task<(IReadOnlyList<CompletionItem>, bool)> BuildCompletionItems(
+            Document document,
+            SourceText sourceText,
+            int position,
+            CSharpCompletionService completionService,
+            CSharpCompletionList completions,
+            TextSpan typedSpan,
+            bool expectingImportedItems,
+            bool isSuggestionMode)
+        {
+            return await BuildCompletionItemsSync(document, sourceText, position, completionService, completions, typedSpan, expectingImportedItems, isSuggestionMode);
+        }
+
+        internal static LinePositionSpanTextChange GetChangeForTextAndSpan(string? insertText, TextSpan changeSpan, SourceText sourceText)
+        {
+            var changeLinePositionSpan = sourceText.Lines.GetLinePositionSpan(changeSpan);
+            return new()
+            {
+                NewText = insertText ?? "",
+                StartLine = changeLinePositionSpan.Start.Line,
+                StartColumn = changeLinePositionSpan.Start.Character,
+                EndLine = changeLinePositionSpan.End.Line,
+                EndColumn = changeLinePositionSpan.End.Character
+            };
+        }
+    }
+}
